+++ InterlocutorComercial/MedidorRN.h
@@ -10,7 +10,7 @@ public:
 	MedidorRN(string nombreArchivo);
 	MedidorRN();
 	~MedidorRN();
-	Medidor BuscarCMedidor(long idMedidor);
+	Medidor BuscarCMedidor(int idMedidor);
 	bool ModificaMedidor(Medidor& medidor);
 	bool AltaMedidor(Medidor& medidor);
 	int IdMedidor(Medidor& medidor);
